requirements_no_versions.txt
amqp
annotated-types
anthropic
anyio
asgiref
asttokens
attrs
backcall
beautifulsoup4
billiard
bleach
cachetools
celery
certifi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
decorator
defusedxml
distro
Django
django-cors-headers
djangorestframework
djangorestframework_simplejwt
docopt
docstring_parser
einops
exceptiongroup
executing
fastjsonschema
filelock
fsspec
google-auth
google-auth-oauthlib
google-genai
h11
httpcore
httpx
idna
ipython
jedi
Jinja2
jiter
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
jupyterlab_pygments
kombu
MarkupSafe
matplotlib-inline
mistune
mpmath
nbclient
nbconvert
nbformat
networkx
numpy
oauthlib
packaging
pandocfilters
parso
pickleshare
pillow
pipreqs
platformdirs
prompt_toolkit
pure_eval
pyasn1
pyasn1_modules
pydantic
pydantic_core
pydicom
Pygments
PyJWT
python-dateutil
python-decouple
pyzmq
referencing
requests
requests-oauthlib
rpds-py
rsa
six
sniffio
soupsieve
sqlparse
stack-data
sympy
tenacity
tinycss2
torch
torchaudio
torchvision
tornado
traitlets
typing-inspection
typing_extensions
tzdata
tzlocal
urllib3
vine
wcwidth
webencodings
websockets
yarg
